+++ src/element_record_column_cuts.rs
@@ -248,6 +248,7 @@ mod tests {
             owner_reference: None,
             references,
             id_from_enclosing_record: false,
+            design_option: None,
         }
     }